% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards-helpers.R
\name{get_standard_names}
\alias{get_standard_names}
\title{Get Standard Names}
\usage{
get_standard_names()
}
\value{
Character vector of canonical standard names.
}
\description{
Returns a character vector of all recognized standard names.
}
\examples{
get_standard_names()

}
\seealso{
Other standards: 
\code{\link{auto_detect_standards}()},
\code{\link{get_recognized_standards}()},
\code{\link{get_standard_type}()},
\code{\link{is_standard}()},
\code{\link{match_standards}()},
\code{\link{normalize_standard_name}()}
}
\concept{standards}
