% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards-helpers.R
\name{get_standard_type}
\alias{get_standard_type}
\title{Get Standard Type}
\usage{
get_standard_type(standard_name)
}
\arguments{
\item{standard_name}{Character string of the standard name (canonical form).}
}
\value{
Character string of the standard type: "PS", "PMMA", "PEG",
"Pullulan", "Generic", or "Other".
}
\description{
Classifies a canonical standard name into its polymer type.
}
\examples{
get_standard_type("PS A")
get_standard_type("Standard 2")
get_standard_type("PMMA C")

}
\seealso{
Other standards: 
\code{\link{auto_detect_standards}()},
\code{\link{get_recognized_standards}()},
\code{\link{get_standard_names}()},
\code{\link{is_standard}()},
\code{\link{match_standards}()},
\code{\link{normalize_standard_name}()}
}
\concept{standards}
