% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.sec-package.R
\docType{package}
\name{measure.sec-package}
\alias{measure.sec}
\alias{measure.sec-package}
\title{measure.sec: SEC/GPC Analysis Extensions for the Measure Package}
\description{
Provides preprocessing and analysis recipe steps for Size Exclusion Chromatography (SEC) and Gel Permeation Chromatography (GPC) data. Includes multi-detector support (RI, UV, MALS), inter-detector delay correction, molecular weight calculations (Mn, Mw, Mz, dispersity), and calibration curve fitting. This package is a technique pack for the 'measure' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JamesHWade/measure-sec}
  \item Report bugs at \url{https://github.com/JamesHWade/measure-sec/issues}
}

}
\author{
\strong{Maintainer}: James Wade \email{github@jameshwade.com} (\href{https://orcid.org/0000-0002-9740-1905}{ORCID})

}
\keyword{internal}
