% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_sec_baseline.R,
%   R/step_sec_concentration.R, R/step_sec_detector_delay.R,
%   R/step_sec_mw_averages.R, R/step_sec_mw_distribution.R,
%   R/step_sec_mw_fractions.R, R/step_sec_ri.R, R/step_sec_uv.R,
%   R/tidy-methods.R
\name{required_pkgs.step_sec_baseline}
\alias{required_pkgs.step_sec_baseline}
\alias{required_pkgs.step_sec_concentration}
\alias{required_pkgs.step_sec_detector_delay}
\alias{required_pkgs.step_sec_mw_averages}
\alias{required_pkgs.step_sec_mw_distribution}
\alias{required_pkgs.step_sec_mw_fractions}
\alias{required_pkgs.step_sec_ri}
\alias{required_pkgs.step_sec_uv}
\alias{required_pkgs.step_sec}
\title{Required Packages for SEC Steps}
\usage{
\method{required_pkgs}{step_sec_baseline}(x, ...)

\method{required_pkgs}{step_sec_concentration}(x, ...)

\method{required_pkgs}{step_sec_detector_delay}(x, ...)

\method{required_pkgs}{step_sec_mw_averages}(x, ...)

\method{required_pkgs}{step_sec_mw_distribution}(x, ...)

\method{required_pkgs}{step_sec_mw_fractions}(x, ...)

\method{required_pkgs}{step_sec_ri}(x, ...)

\method{required_pkgs}{step_sec_uv}(x, ...)
}
\arguments{
\item{x}{A recipe step object.}

\item{...}{Not currently used.}
}
\value{
A character vector of package names.
}
\description{
Returns the packages required to use the step.
}
\keyword{internal}
