% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_sec_baseline.R, R/step_sec_mw_averages.R,
%   R/step_sec_mw_distribution.R, R/step_sec_mw_fractions.R, R/tidy-methods.R
\name{tidy.step_sec_baseline}
\alias{tidy.step_sec_baseline}
\alias{tidy.step_sec_mw_averages}
\alias{tidy.step_sec_mw_distribution}
\alias{tidy.step_sec_mw_fractions}
\alias{tidy.step_sec}
\title{Tidy Methods for SEC Steps}
\usage{
\method{tidy}{step_sec_baseline}(x, ...)

\method{tidy}{step_sec_mw_averages}(x, ...)

\method{tidy}{step_sec_mw_distribution}(x, ...)

\method{tidy}{step_sec_mw_fractions}(x, ...)
}
\arguments{
\item{x}{A recipe step object.}

\item{...}{Not currently used.}
}
\value{
A tibble with step information.
}
\description{
The \code{tidy()} method returns a tibble with information about the step
parameters.
}
\keyword{internal}
