% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_sec_aggregates.R,
%   R/step_sec_band_broadening.R, R/step_sec_baseline.R,
%   R/step_sec_broad_standard.R, R/step_sec_composition.R,
%   R/step_sec_concentration.R, R/step_sec_conventional_cal.R,
%   R/step_sec_dad.R, R/step_sec_detector_delay.R, R/step_sec_dls.R,
%   R/step_sec_integration_window.R, R/step_sec_intrinsic_visc.R,
%   R/step_sec_lals.R, R/step_sec_mals.R, R/step_sec_mw_averages.R,
%   R/step_sec_mw_distribution.R, R/step_sec_mw_fractions.R,
%   R/step_sec_oligomer.R, R/step_sec_peaks_detect.R, R/step_sec_protein.R,
%   R/step_sec_rals.R, R/step_sec_ri.R, R/step_sec_universal_cal.R,
%   R/step_sec_uv.R, R/step_sec_uv_ri_ratio.R, R/step_sec_viscometer.R,
%   R/tidy-methods.R
\name{required_pkgs.step_sec_aggregates}
\alias{required_pkgs.step_sec_aggregates}
\alias{required_pkgs.step_sec_band_broadening}
\alias{required_pkgs.step_sec_baseline}
\alias{required_pkgs.step_sec_broad_standard}
\alias{required_pkgs.step_sec_composition}
\alias{required_pkgs.step_sec_concentration}
\alias{required_pkgs.step_sec_conventional_cal}
\alias{required_pkgs.step_sec_dad}
\alias{required_pkgs.step_sec_detector_delay}
\alias{required_pkgs.step_sec_dls}
\alias{required_pkgs.step_sec_integration_window}
\alias{required_pkgs.step_sec_intrinsic_visc}
\alias{required_pkgs.step_sec_lals}
\alias{required_pkgs.step_sec_mals}
\alias{required_pkgs.step_sec_mw_averages}
\alias{required_pkgs.step_sec_mw_distribution}
\alias{required_pkgs.step_sec_mw_fractions}
\alias{required_pkgs.step_sec_oligomer}
\alias{required_pkgs.step_sec_peaks_detect}
\alias{required_pkgs.step_sec_protein}
\alias{required_pkgs.step_sec_rals}
\alias{required_pkgs.step_sec_ri}
\alias{required_pkgs.step_sec_universal_cal}
\alias{required_pkgs.step_sec_uv}
\alias{required_pkgs.step_sec_uv_ri_ratio}
\alias{required_pkgs.step_sec_viscometer}
\alias{required_pkgs.step_sec}
\title{Required Packages for SEC Steps}
\usage{
\method{required_pkgs}{step_sec_aggregates}(x, ...)

\method{required_pkgs}{step_sec_band_broadening}(x, ...)

\method{required_pkgs}{step_sec_baseline}(x, ...)

\method{required_pkgs}{step_sec_broad_standard}(x, ...)

\method{required_pkgs}{step_sec_composition}(x, ...)

\method{required_pkgs}{step_sec_concentration}(x, ...)

\method{required_pkgs}{step_sec_conventional_cal}(x, ...)

\method{required_pkgs}{step_sec_dad}(x, ...)

\method{required_pkgs}{step_sec_detector_delay}(x, ...)

\method{required_pkgs}{step_sec_dls}(x, ...)

\method{required_pkgs}{step_sec_integration_window}(x, ...)

\method{required_pkgs}{step_sec_intrinsic_visc}(x, ...)

\method{required_pkgs}{step_sec_lals}(x, ...)

\method{required_pkgs}{step_sec_mals}(x, ...)

\method{required_pkgs}{step_sec_mw_averages}(x, ...)

\method{required_pkgs}{step_sec_mw_distribution}(x, ...)

\method{required_pkgs}{step_sec_mw_fractions}(x, ...)

\method{required_pkgs}{step_sec_oligomer}(x, ...)

\method{required_pkgs}{step_sec_peaks_detect}(x, ...)

\method{required_pkgs}{step_sec_protein}(x, ...)

\method{required_pkgs}{step_sec_rals}(x, ...)

\method{required_pkgs}{step_sec_ri}(x, ...)

\method{required_pkgs}{step_sec_universal_cal}(x, ...)

\method{required_pkgs}{step_sec_uv}(x, ...)

\method{required_pkgs}{step_sec_uv_ri_ratio}(x, ...)

\method{required_pkgs}{step_sec_viscometer}(x, ...)
}
\arguments{
\item{x}{A recipe step object.}

\item{...}{Not currently used.}
}
\value{
A character vector of package names.
}
\description{
Returns the packages required to use the step.
}
\keyword{internal}
