% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_sec_aggregates.R, R/step_sec_baseline.R,
%   R/step_sec_composition.R, R/step_sec_concentration.R,
%   R/step_sec_detector_delay.R, R/step_sec_intrinsic_visc.R,
%   R/step_sec_mals.R, R/step_sec_mw_averages.R, R/step_sec_mw_distribution.R,
%   R/step_sec_mw_fractions.R, R/step_sec_ri.R, R/step_sec_uv.R,
%   R/step_sec_uv_ri_ratio.R, R/step_sec_viscometer.R, R/tidy-methods.R
\name{tidy.step_sec_aggregates}
\alias{tidy.step_sec_aggregates}
\alias{tidy.step_sec_baseline}
\alias{tidy.step_sec_composition}
\alias{tidy.step_sec_concentration}
\alias{tidy.step_sec_detector_delay}
\alias{tidy.step_sec_intrinsic_visc}
\alias{tidy.step_sec_mals}
\alias{tidy.step_sec_mw_averages}
\alias{tidy.step_sec_mw_distribution}
\alias{tidy.step_sec_mw_fractions}
\alias{tidy.step_sec_ri}
\alias{tidy.step_sec_uv}
\alias{tidy.step_sec_uv_ri_ratio}
\alias{tidy.step_sec_viscometer}
\alias{tidy.step_sec}
\title{Tidy Methods for SEC Steps}
\usage{
\method{tidy}{step_sec_aggregates}(x, ...)

\method{tidy}{step_sec_baseline}(x, ...)

\method{tidy}{step_sec_composition}(x, ...)

\method{tidy}{step_sec_concentration}(x, ...)

\method{tidy}{step_sec_detector_delay}(x, ...)

\method{tidy}{step_sec_intrinsic_visc}(x, ...)

\method{tidy}{step_sec_mals}(x, ...)

\method{tidy}{step_sec_mw_averages}(x, ...)

\method{tidy}{step_sec_mw_distribution}(x, ...)

\method{tidy}{step_sec_mw_fractions}(x, ...)

\method{tidy}{step_sec_ri}(x, ...)

\method{tidy}{step_sec_uv}(x, ...)

\method{tidy}{step_sec_uv_ri_ratio}(x, ...)

\method{tidy}{step_sec_viscometer}(x, ...)
}
\arguments{
\item{x}{A recipe step object.}

\item{...}{Not currently used.}
}
\value{
A tibble with step information.
}
\description{
The \code{tidy()} method returns a tibble with information about the step
parameters.
}
\keyword{internal}
